refactor(cli): rename schema add to collection add (#597)
Port Go DefraDB #4591. Remove standalone `schema` CLI command and make
it an alias for `collection`. The `schema add` functionality was already
duplicated in `collection add`. Move `schema describe` to `collection
schema` subcommand.
